requirements.txt
pynetbox==7.0.1
python-dotenv==1.0.0
